requirements.txt
pytorch
matplotlib
tqdm
yacs
Pillow
tensorboardx
dill
wandb
scikit-learn
opencv-python
coloredlogs